Move Menu to its own file.
